+++ Managers/FrameMan.h
@@ -201,6 +201,15 @@ namespace RTE {
 		int CalculateTextHeight(const std::string &text, int maxWidth, bool isSmall);
 
 		/// <summary>
+		/// Gets a copy of the passed in string, split into multiple lines as needed to fit within the specified width limit, based on the font to use.
+		/// </summary>
+		/// <param name="stringToSplit">The string to get a split copy of.</param>
+		/// <param name="widthLimit">The maximum width each line of the string can be.</param>
+		/// <param name="fontToUse">The font the string will use for calculating the string's width.</param>
+		/// <returns>A copy of the passed in string, split into multiple lines as needed.</returns>
+		std::string SplitStringToFitWidth(const std::string &stringToSplit, int widthLimit, bool useSmallFont);
+
+		/// <summary>
 		/// Gets the message to be displayed on top of each player's screen.
 		/// </summary>
 		/// <param name="whichScreen">Which player screen to get message from.</param>
@@ -503,7 +512,7 @@ namespace RTE {
 		std::unique_ptr<BITMAP, BitmapDeleter> m_BackBuffer8; //!< Screen backbuffer, always 8bpp, gets copied to the 32bpp buffer for post-processing.
 		std::unique_ptr<BITMAP, BitmapDeleter> m_BackBuffer32; //!< 32bpp backbuffer, only used for post-processing.
 		std::unique_ptr<BITMAP, BitmapDeleter> m_OverlayBitmap32; //!< 32bpp bitmap used for overlaying (fading in/out or darkening) the screen.
-		std::unique_ptr<BITMAP, BitmapDeleter> m_ScreenDumpBuffer; //!< Temporary buffer for making quick screencaps.
+		std::unique_ptr<BITMAP, BitmapDeleter> m_ScreenDumpBuffer; //!< Temporary buffer for making quick screencaps. This is used for color conversion between 32bpp and 24bpp so we can save the file.
 		std::unique_ptr<BITMAP, BitmapDeleter> m_WorldDumpBuffer; //!< Temporary buffer for making whole scene screencaps.
 		std::unique_ptr<BITMAP, BitmapDeleter> m_ScenePreviewDumpGradient; //!< BITMAP for the scene preview sky gradient (easier to load from a pre-made file because it's dithered).
 		std::unique_ptr<BITMAP, BitmapDeleter> m_ScreenDumpNamePlaceholder; //!< Dummy BITMAP for keeping naming continuity when saving ScreenDumps with multi-threading.