+++ Menus/TitleScreen.cpp
@@ -17,6 +17,8 @@ namespace RTE {
 	void TitleScreen::Clear() {
 		m_FadeAmount = 0;
 
+		m_TitleScreenMaxWidth = 0;
+
 		m_BackdropStars.clear();
 		m_Nebula.Reset();
 		m_Moon.Reset();
@@ -69,6 +71,8 @@ namespace RTE {
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 	void TitleScreen::Create(AllegroScreen *guiScreen) {
+		m_TitleScreenMaxWidth = g_WindowMan.FullyCoversAllDisplays() ? g_WindowMan.GetPrimaryWindowDisplayWidth() / g_WindowMan.GetResMultiplier() : g_WindowMan.GetResX();
+
 		CreateTitleElements();
 
 		m_IntroScrollStartOffsetY = (static_cast<float>(m_Nebula.GetBitmap()->h) / m_Nebula.GetScrollRatio().GetY()) - (static_cast<float>(g_WindowMan.GetResY()) / m_Nebula.GetScrollRatio().GetY());
@@ -97,7 +101,11 @@ namespace RTE {
 		m_FmodLogo = ContentFile("Base.rte/GUIs/Title/Intro/FMODLogo.png").GetAsBitmap();
 
 		m_PreGameLogoText.Create(ContentFile("Base.rte/GUIs/Title/Intro/PreTitle.png"));
+		m_PreGameLogoText.SetPos(Vector(static_cast<float>(m_TitleScreenMaxWidth / 2), static_cast<float>(g_WindowMan.GetResY() / 2)));
+
 		m_PreGameLogoTextGlow.Create(ContentFile("Base.rte/GUIs/Title/Intro/PreTitleGlow.png"));
+		m_PreGameLogoTextGlow.SetPos(m_PreGameLogoText.GetPos());
+
 		m_GameLogo.Create(ContentFile("Base.rte/GUIs/Title/Title.png"));
 		m_GameLogoGlow.Create(ContentFile("Base.rte/GUIs/Title/TitleGlow.png"));
 		m_Planet.Create(ContentFile("Base.rte/GUIs/Title/Planet.png"));
@@ -231,7 +239,7 @@ namespace RTE {
 		if (skipSlideshow && (m_IntroSequenceState > IntroSequence::SlideshowFadeIn && m_IntroSequenceState != IntroSequence::MainMenuAppear)) {
 			m_SectionSwitch = true;
 			m_IntroSequenceState = IntroSequence::MainMenuAppear;
-			m_GameLogo.SetPos(Vector(static_cast<float>(g_WindowMan.GetResX() / 2), 120));
+			m_GameLogo.SetPos(Vector(static_cast<float>(m_TitleScreenMaxWidth / 2), 120));
 			m_StationOrbitTimer.SetElapsedRealTimeS(19);
 			return;
 		}
@@ -331,9 +339,7 @@ namespace RTE {
 			case IntroSequence::SlideshowEnd:
 				if (m_SectionSwitch) {
 					SetSectionDurationAndResetSwitch(66.6F - static_cast<float>(m_IntroSongTimer.GetElapsedRealTimeS()));
-					m_PreGameLogoText.SetPos(Vector(static_cast<float>(g_WindowMan.GetResX() / 2), static_cast<float>(g_WindowMan.GetResY() / 2)));
-					m_PreGameLogoTextGlow.SetPos(m_PreGameLogoText.GetPos());
-					m_GameLogo.SetPos(Vector(static_cast<float>(g_WindowMan.GetResX() / 2), static_cast<float>(g_WindowMan.GetResY() / 2) - 20));
+					m_GameLogo.SetPos(Vector(static_cast<float>(m_TitleScreenMaxWidth / 2), static_cast<float>(g_WindowMan.GetResY() / 2) - 20));
 				}
 				break;
 			default:
@@ -360,7 +366,7 @@ namespace RTE {
 					SetSectionDurationAndResetSwitch(92.4F - static_cast<float>(m_IntroSongTimer.GetElapsedRealTimeS()));
 					clear_to_color(g_FrameMan.GetOverlayBitmap32(), 0);
 				}
-				if (m_SectionProgress > 0.5F) { m_GameLogo.SetPos(Vector(static_cast<float>(g_WindowMan.GetResX() / 2), EaseIn((static_cast<float>(g_WindowMan.GetResY() / 2)) - 20, 120, (m_SectionProgress - 0.5F) / 0.5F))); }
+				if (m_SectionProgress > 0.5F) { m_GameLogo.SetPos(Vector(static_cast<float>(m_TitleScreenMaxWidth / 2), EaseIn((static_cast<float>(g_WindowMan.GetResY() / 2)) - 20, 120, (m_SectionProgress - 0.5F) / 0.5F))); }
 				break;
 			case IntroSequence::PreMainMenu:
 				if (m_SectionSwitch) {
@@ -375,7 +381,7 @@ namespace RTE {
 					g_AudioMan.PlayMusic("Base.rte/Music/Hubnester/ccmenu.ogg", -1);
 				}
 				m_ScrollOffset.SetY(EaseOut(m_PreMainMenuScrollOffsetY, 0, m_SectionProgress));
-				m_GameLogo.SetPos(Vector(static_cast<float>(g_WindowMan.GetResX() / 2), EaseOut(120, m_GameLogoMainMenuOffsetY, m_SectionProgress)));
+				m_GameLogo.SetPos(Vector(static_cast<float>(m_TitleScreenMaxWidth / 2), EaseOut(120, m_GameLogoMainMenuOffsetY, m_SectionProgress)));
 				if (m_SectionElapsedTime >= m_SectionDuration) {
 					SetTitleTransitionState(TitleTransition::MainMenu);
 					m_FinishedPlayingIntro = true;
@@ -394,7 +400,7 @@ namespace RTE {
 				if (m_SectionSwitch) {
 					m_SectionSwitch = false;
 					m_ScrollOffset.SetY(0);
-					m_GameLogo.SetPos(Vector(static_cast<float>(g_WindowMan.GetResX() / 2), m_GameLogoMainMenuOffsetY));
+					m_GameLogo.SetPos(Vector(static_cast<float>(m_TitleScreenMaxWidth / 2), m_GameLogoMainMenuOffsetY));
 				}
 				break;
 			case TitleTransition::ScenarioMenu:
@@ -412,7 +418,7 @@ namespace RTE {
 					g_AudioMan.PlayMusic("Base.rte/Music/dBSoundworks/thisworld5.ogg", -1);
 				}
 				m_ScrollOffset.SetY(EaseOut(0, m_PlanetViewScrollOffsetY, m_SectionProgress));
-				m_GameLogo.SetPos(Vector(static_cast<float>(g_WindowMan.GetResX() / 2), EaseOut(m_GameLogoMainMenuOffsetY, m_GameLogoPlanetViewOffsetY, m_SectionProgress)));
+				m_GameLogo.SetPos(Vector(static_cast<float>(m_TitleScreenMaxWidth / 2), EaseOut(m_GameLogoMainMenuOffsetY, m_GameLogoPlanetViewOffsetY, m_SectionProgress)));
 				if (m_SectionElapsedTime >= m_SectionDuration) { SetTitleTransitionState((m_TitleTransitionState == TitleTransition::MainMenuToScenario) ? TitleTransition::ScenarioMenu : TitleTransition::MetaGameMenu); }
 				break;
 			case TitleTransition::PlanetToMainMenu:
@@ -421,7 +427,7 @@ namespace RTE {
 					g_AudioMan.PlayMusic("Base.rte/Music/Hubnester/ccmenu.ogg", -1);
 				}
 				m_ScrollOffset.SetY(EaseOut(m_PlanetViewScrollOffsetY, 0, m_SectionProgress));
-				m_GameLogo.SetPos(Vector(static_cast<float>(g_WindowMan.GetResX() / 2), EaseOut(m_GameLogoPlanetViewOffsetY, m_GameLogoMainMenuOffsetY, m_SectionProgress)));
+				m_GameLogo.SetPos(Vector(static_cast<float>(m_TitleScreenMaxWidth / 2), EaseOut(m_GameLogoPlanetViewOffsetY, m_GameLogoMainMenuOffsetY, m_SectionProgress)));
 				if (m_SectionElapsedTime >= m_SectionDuration) { SetTitleTransitionState(TitleTransition::MainMenu); }
 				break;
 			case TitleTransition::MainMenuToCredits:
@@ -440,7 +446,7 @@ namespace RTE {
 				if (m_SectionSwitch) {
 					SetSectionDurationAndResetSwitch(0.75F * g_SettingsMan.GetMenuTransitionDurationMultiplier());
 					m_ScrollOffset.SetY(m_PlanetViewScrollOffsetY);
-					m_GameLogo.SetPos(Vector(static_cast<float>(g_WindowMan.GetResX() / 2), m_GameLogoPlanetViewOffsetY));
+					m_GameLogo.SetPos(Vector(static_cast<float>(m_TitleScreenMaxWidth / 2), m_GameLogoPlanetViewOffsetY));
 					m_StationOrbitTimer.SetElapsedRealTimeS(m_StationOrbitTimerElapsedTime);
 					g_AudioMan.PlayMusic("Base.rte/Music/dBSoundworks/thisworld5.ogg", -1);
 				}
@@ -462,7 +468,7 @@ namespace RTE {
 				}
 				g_AudioMan.SetTempMusicVolume(EaseOut(0, 1.0F, m_SectionProgress));
 				m_ScrollOffset.SetY(EaseOut(250, 0, m_SectionProgress));
-				m_GameLogo.SetPos(Vector(static_cast<float>(g_WindowMan.GetResX() / 2), EaseOut(m_GameLogoPlanetViewOffsetY, m_GameLogoMainMenuOffsetY, m_SectionProgress)));
+				m_GameLogo.SetPos(Vector(static_cast<float>(m_TitleScreenMaxWidth / 2), EaseOut(m_GameLogoPlanetViewOffsetY, m_GameLogoMainMenuOffsetY, m_SectionProgress)));
 				m_FadeAmount = static_cast<int>(EaseOut(255, 0, m_SectionProgress));
 				if (m_SectionElapsedTime >= m_SectionDuration) { SetTitleTransitionState(TitleTransition::MainMenu); }
 				break;
@@ -471,7 +477,7 @@ namespace RTE {
 				if (m_SectionSwitch) { SetSectionDurationAndResetSwitch(0.75F * g_SettingsMan.GetMenuTransitionDurationMultiplier()); }
 				g_AudioMan.SetTempMusicVolume(EaseIn(1.0F, 0, m_SectionProgress));
 				m_ScrollOffset.SetY(EaseIn(0, 250, m_SectionProgress));
-				m_GameLogo.SetPos(Vector(static_cast<float>(g_WindowMan.GetResX() / 2), EaseIn(m_GameLogoMainMenuOffsetY, m_GameLogoPlanetViewOffsetY, m_SectionProgress)));
+				m_GameLogo.SetPos(Vector(static_cast<float>(m_TitleScreenMaxWidth / 2), EaseIn(m_GameLogoMainMenuOffsetY, m_GameLogoPlanetViewOffsetY, m_SectionProgress)));
 				m_FadeAmount = static_cast<int>(EaseIn(0, 255, m_SectionProgress));
 				if (m_SectionElapsedTime >= m_SectionDuration) { SetTitleTransitionState((m_TitleTransitionState == TitleTransition::ScrollingFadeOutQuit) ? TitleTransition::TransitionEndQuit : TitleTransition::TransitionEnd); }
 				break;
@@ -494,15 +500,15 @@ namespace RTE {
 			if (m_IntroSequenceState >= IntroSequence::GameLogoAppear) { DrawGameLogo(); }
 
 			if (m_IntroSequenceState >= IntroSequence::DataRealmsLogoFadeIn && m_IntroSequenceState <= IntroSequence::DataRealmsLogoFadeOut) {
-				draw_sprite(g_FrameMan.GetBackBuffer32(), m_DataRealmsLogo, (g_WindowMan.GetResX() - m_DataRealmsLogo->w) / 2, (g_WindowMan.GetResY() - m_DataRealmsLogo->h) / 2);
+				draw_sprite(g_FrameMan.GetBackBuffer32(), m_DataRealmsLogo, (m_TitleScreenMaxWidth - m_DataRealmsLogo->w) / 2, (g_WindowMan.GetResY() - m_DataRealmsLogo->h) / 2);
 				std::string copyrightNotice(64, '\0');
 				std::snprintf(copyrightNotice.data(), copyrightNotice.size(), "Cortex Command is TM and %c 2017 Data Realms, LLC", -35);
 				AllegroBitmap guiBackBuffer(g_FrameMan.GetBackBuffer32());
-				m_IntroTextFont->DrawAligned(&guiBackBuffer, g_WindowMan.GetResX() / 2, g_WindowMan.GetResY() - m_IntroTextFont->GetFontHeight() - 5, copyrightNotice, GUIFont::Centre);
+				m_IntroTextFont->DrawAligned(&guiBackBuffer, m_TitleScreenMaxWidth / 2, g_WindowMan.GetResY() - m_IntroTextFont->GetFontHeight() - 5, copyrightNotice, GUIFont::Centre);
 			} else if (m_IntroSequenceState >= IntroSequence::FmodLogoFadeIn && m_IntroSequenceState <= IntroSequence::FmodLogoFadeOut) {
-				draw_sprite(g_FrameMan.GetBackBuffer32(), m_FmodLogo, (g_WindowMan.GetResX() - m_FmodLogo->w) / 2, (g_WindowMan.GetResY() - m_FmodLogo->h) / 2);
+				draw_sprite(g_FrameMan.GetBackBuffer32(), m_FmodLogo, (m_TitleScreenMaxWidth - m_FmodLogo->w) / 2, (g_WindowMan.GetResY() - m_FmodLogo->h) / 2);
 				AllegroBitmap guiBackBuffer(g_FrameMan.GetBackBuffer32());
-				m_IntroTextFont->DrawAligned(&guiBackBuffer, g_WindowMan.GetResX() / 2, g_WindowMan.GetResY() - m_IntroTextFont->GetFontHeight() - 5, "Made with FMOD Studio by Firelight Technologies Pty Ltd.", GUIFont::Centre);
+				m_IntroTextFont->DrawAligned(&guiBackBuffer, m_TitleScreenMaxWidth / 2, g_WindowMan.GetResY() - m_IntroTextFont->GetFontHeight() - 5, "Made with FMOD Studio by Firelight Technologies Pty Ltd.", GUIFont::Centre);
 			} else if (m_IntroSequenceState >= IntroSequence::ShowSlide1 && m_IntroSequenceState <= IntroSequence::ShowSlide8) {
 				DrawSlideshowSlide();
 			} else if (m_IntroSequenceState == IntroSequence::SlideshowEnd) {
@@ -529,7 +535,7 @@ namespace RTE {
 
 	void TitleScreen::DrawTitleScreenScene() {
 		Box nebulaTargetBox;
-		m_Nebula.SetOffset(Vector(static_cast<float>((g_WindowMan.GetResX() - m_Nebula.GetBitmap()->w) / 2), m_ScrollOffset.GetY()));
+		m_Nebula.SetOffset(Vector(static_cast<float>((m_TitleScreenMaxWidth - m_Nebula.GetBitmap()->w) / 2), m_ScrollOffset.GetY()));
 		m_Nebula.Draw(g_FrameMan.GetBackBuffer32(), nebulaTargetBox, true);
 
 		for (const Star &star : m_BackdropStars) {
@@ -539,7 +545,7 @@ namespace RTE {
 			draw_trans_sprite(g_FrameMan.GetBackBuffer32(), star.Bitmap, star.Position.GetFloorIntX(), starPosY);
 		}
 
-		m_PlanetPos.SetXY(static_cast<float>(g_WindowMan.GetResX() / 2), static_cast<float>(567 - m_ScrollOffset.GetFloorIntY()));
+		m_PlanetPos.SetXY(static_cast<float>(m_TitleScreenMaxWidth / 2), static_cast<float>(567 - m_ScrollOffset.GetFloorIntY()));
 		m_Moon.SetPos(Vector(m_PlanetPos.GetX() + 200, 364 - (m_ScrollOffset.GetY() * 0.60F)));
 		m_Moon.Draw(g_FrameMan.GetBackBuffer32(), Vector(), DrawMode::g_DrawAlpha);
 		m_Planet.SetPos(m_PlanetPos);
@@ -571,24 +577,31 @@ namespace RTE {
 
 	void TitleScreen::DrawSlideshowSlide() {
 		int slide = static_cast<int>(m_IntroSequenceState) - static_cast<int>(IntroSequence::ShowSlide1);
-		Vector slidePos(static_cast<float>((g_WindowMan.GetResX() / 2) - (m_IntroSlides.at(slide)->w / 2)), static_cast<float>((g_WindowMan.GetResY() / 2) - (m_IntroSlides.at(slide)->h / 2)));
+		Vector slidePos(static_cast<float>((m_TitleScreenMaxWidth - m_IntroSlides.at(slide)->w) / 2), static_cast<float>((g_WindowMan.GetResY() / 2) - (m_IntroSlides.at(slide)->h / 2)));
 
-		if (m_IntroSlides.at(slide)->w > g_WindowMan.GetResX()) {
+		if (m_IntroSlides.at(slide)->w > m_TitleScreenMaxWidth) {
 			if (m_SectionElapsedTime < m_SlideFadeInDuration) {
 				slidePos.SetX(0);
 			} else if (m_SectionElapsedTime < m_SectionDuration - m_SlideFadeOutDuration) {
-				slidePos.SetX(EaseInOut(0, static_cast<float>(g_WindowMan.GetResX() - m_IntroSlides.at(slide)->w), (m_SectionElapsedTime - m_SlideFadeInDuration) / (m_SectionDuration - m_SlideFadeInDuration - m_SlideFadeOutDuration)));
+				slidePos.SetX(EaseInOut(0, static_cast<float>(m_TitleScreenMaxWidth - m_IntroSlides.at(slide)->w), (m_SectionElapsedTime - m_SlideFadeInDuration) / (m_SectionDuration - m_SlideFadeInDuration - m_SlideFadeOutDuration)));
 			} else {
-				slidePos.SetX(static_cast<float>(g_WindowMan.GetResX() - m_IntroSlides.at(slide)->w));
+				slidePos.SetX(static_cast<float>(m_TitleScreenMaxWidth - m_IntroSlides.at(slide)->w));
 			}
 		}
 		int fadeAmount = static_cast<int>((m_SectionElapsedTime < m_SlideFadeInDuration) ? EaseOut(0, 255.0F, m_SectionElapsedTime / m_SlideFadeInDuration) : EaseIn(255.0F, 0, (m_SectionElapsedTime - m_SectionDuration + m_SlideFadeOutDuration) / m_SlideFadeOutDuration));
 		set_trans_blender(fadeAmount, fadeAmount, fadeAmount, fadeAmount);
+
+		// Set a clipping rect so parts of slides that are too wide to fit the primary window's display (left-most) don't get drawn on other displays. This only has effect in multi-display fullscreen.
+		set_clip_rect(g_FrameMan.GetBackBuffer32(), 0, 0, m_TitleScreenMaxWidth - g_WindowMan.GetResMultiplier(), g_WindowMan.GetResY());
+
 		draw_trans_sprite(g_FrameMan.GetBackBuffer32(), m_IntroSlides.at(slide), slidePos.GetFloorIntX(), slidePos.GetFloorIntY());
+
+		// Have to immediately reset the clipping rect otherwise the stars in the other displays slowly go into warp speed until the intro sequence is done.
+		set_clip_rect(g_FrameMan.GetBackBuffer32(), 0, 0, g_FrameMan.GetBackBuffer32()->w, g_FrameMan.GetBackBuffer32()->h);
 
 		if (!m_SlideshowSlideText.empty()) {
 			AllegroBitmap guiBackBuffer(g_FrameMan.GetBackBuffer32());
-			m_IntroTextFont->DrawAligned(&guiBackBuffer, g_WindowMan.GetResX() / 2, (g_WindowMan.GetResY() / 2) + (m_IntroSlides.at(slide)->h / 2) + 12, m_SlideshowSlideText, GUIFont::Centre);
+			m_IntroTextFont->DrawAligned(&guiBackBuffer, m_TitleScreenMaxWidth / 2, (g_WindowMan.GetResY() / 2) + (m_IntroSlides.at(slide)->h / 2) + 12, m_SlideshowSlideText, GUIFont::Centre);
 		}
 	}